Add SerializedValue.hasTag replacing MessageSerialization.makeSimpleName
Instead of duplicating the type discriminator out-of-band, add a
way to test the usual type discriminator. This is obviously not
compatible with duck typing.

We don't want to just use SerializedValue.parse[T] because failed
matches would incur exception-throwing overhead, and because we
really don't want duck typing in this context because we might
accidentally match the wrong kind of event.